requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
openai==1.52.2
httpx==0.27.2
faiss-cpu==1.8.0.post1
numpy==1.26.4
pandas==2.2.3
tqdm==4.66.5
